refactor _calc_score
